final app/backend/requirements2.txt
aenum==3.1.16
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
bcrypt==3.2.2
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
dnspython==2.8.0
earthengine-api==1.6.8
email-validator==2.3.0
fastapi==0.116.1
filelock==3.19.1
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
GeoAlchemy2==0.18.0
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
h11==0.16.0
httplib2==0.31.0
httptools==0.6.4
idna==3.10
imagecodecs==2025.8.2
imageio==2.37.0
inference-sdk==0.56.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
labelImg==1.8.6
lazy_loader==0.4
lxml==6.0.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.5
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.10.0.84
packaging==25.0
passlib==1.7.4
pillow==11.3.0
polars==1.33.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.4
pyproj==3.7.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
rasterio==1.4.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
sentinelhub==3.11.2
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
supervision==0.26.1
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.9.9
tomli==2.2.1
tomli_w==1.2.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
ultralytics==8.3.203
ultralytics-thop==2.0.17
uritemplate==4.2.0
urllib3==2.5.0
utm==0.8.1
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
